Split GUI into common lib, cli, and gui versions
