+++ app/OTMonsterConverter/App.xaml
@@ -1,0 +1,9 @@
+﻿<Application x:Class="OTMonsterConverter2.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:OTMonsterConverter"
+             StartupUri="MainWindow.xaml">
+    <Application.Resources>
+         
+    </Application.Resources>
+</Application>
