+++ app/OTMonsterConverter/ConvertErrorImageToImageFileConverter.cs
@@ -1,10 +1,6 @@
 ﻿using MonsterConverterInterface;
 using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Data;
 
 namespace OTMonsterConverter
