+++ app/OTMonsterConverter/MainWindow.xaml.cs
@@ -1,19 +1,10 @@
 ﻿using MonsterConverterInterface;
-using System;
-using System.Collections.Generic;
+using MonsterConverterProcessor;
 using System.Data;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace OTMonsterConverter
 {
