+++ app/otmc/otmc.csproj
@@ -1,11 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net5.0-windows</TargetFramework>
-    <UseWPF>true</UseWPF>
-    <useWindowsForms>true</useWindowsForms>
-    <ApplicationIcon>app.ico</ApplicationIcon>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
     <Authors>Soul4Soul</Authors>
     <AssemblyVersion>3.0.0.0</AssemblyVersion>
   </PropertyGroup>
@@ -21,31 +18,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="app.ico" />
-    <None Remove="Images\error.png" />
-    <None Remove="Images\info.png" />
-    <None Remove="Images\success.png" />
-    <None Remove="Images\warning.png" />
+    <PackageReference Include="Mono.Options" Version="6.6.0.161" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\MonsterConverterInterface\MonsterConverterInterface.csproj" />
     <ProjectReference Include="..\MonsterConverterProcessor\MonsterConverterProcessor.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Resource Include="app.ico" />
-    <Resource Include="Images\error.png" />
-    <Resource Include="Images\info.png" />
-    <Resource Include="Images\success.png" />
-    <Resource Include="Images\warning.png" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Page Update="MainWindow.xaml">
-      <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
   </ItemGroup>
 
 </Project>
